moved button styles to forced color stylesheet behavior
