+++ packages/web-components/fast-components/src/button/button.styles.ts
@@ -2,6 +2,8 @@ import { css } from "@microsoft/fast-element";
 import {
     AccentButtonStyles,
     BaseButtonStyles,
+    disabledCursor,
+    focusVisible,
     HypertextStyles,
     LightweightButtonStyles,
     OutlineButtonStyles,
@@ -29,6 +31,8 @@ import {
     neutralOutlineHoverBehavior,
     neutralOutlineRestBehavior,
 } from "../styles/recipes";
+import { forcedColorsStylesheetBehavior } from "../styles/match-media-stylesheet-behavior";
+import { SystemColors } from "../styles/system-colors";
 
 export const ButtonStyles = css`
     ${BaseButtonStyles}
@@ -57,5 +61,79 @@ export const ButtonStyles = css`
     neutralForegroundRestBehavior,
     neutralOutlineActiveBehavior,
     neutralOutlineHoverBehavior,
-    neutralOutlineRestBehavior
+    neutralOutlineRestBehavior,
+    forcedColorsStylesheetBehavior(
+        css`
+            :host(.disabled),
+            :host(.disabled) .control {
+                forced-color-adjust: none;
+                background: ${SystemColors.ButtonFace};
+                border-color: ${SystemColors.GrayText};
+                color: ${SystemColors.GrayText};
+                cursor: ${disabledCursor};
+                opacity: 1;
+            }
+            :host(.accent) .control {
+                forced-color-adjust: none;
+                background: ${SystemColors.Highlight};
+                color: ${SystemColors.HighlightText};
+            }
+    
+            :host(.accent) .control:hover {
+                background: ${SystemColors.HighlightText};
+                border-color: ${SystemColors.Highlight};
+                color: ${SystemColors.Highlight};
+            }
+    
+            :host(.accent:${focusVisible}) .control {
+                border-color: ${SystemColors.ButtonText};
+                box-shadow: 0 0 0 2px ${SystemColors.HighlightText} inset;
+            }
+    
+            :host(.accent.disabled) .control,
+            :host(.accent.disabled) .control:hover {
+                background: ${SystemColors.ButtonFace};
+                border-color: ${SystemColors.GrayText};
+                color: ${SystemColors.GrayText};
+            }
+            :host(.lightweight) .control:hover {
+                forced-color-adjust: none;
+                color: ${SystemColors.Highlight};
+            }
+    
+            :host(.lightweight) .control:hover .content::before {
+                background: ${SystemColors.Highlight};
+            }
+    
+            :host(.lightweight.disabled) .control {
+                forced-color-adjust: none;
+                color: ${SystemColors.GrayText};
+            }
+        
+            :host(.lightweight.disabled) .control:hover .content::before {
+                background: none;
+            }
+            :host(.outline.disabled) .control {
+                border-color: ${SystemColors.GrayText};
+            }
+            :host(.stealth) .control {
+                forced-color-adjust: none;
+                background: none;
+                border-color: transparent;
+                color: ${SystemColors.ButtonText};
+                fill: currentColor;
+            }
+            :host(.stealth) .control:hover,
+            :host(.stealth:${focusVisible}) .control {
+                background: ${SystemColors.Highlight};
+                border-color: ${SystemColors.Highlight};
+                color: ${SystemColors.HighlightText};
+            }
+            :host(.stealth.disabled) .control {
+                background: none;
+                border-color: transparent;
+                color: ${SystemColors.GrayText};
+            }
+        `
+    )
 );
