+++ packages/web-components/fast-components/src/styles/patterns/button.ts
@@ -2,8 +2,6 @@ import { css } from "@microsoft/fast-element";
 import { display } from "../display";
 import { focusVisible } from "../focus";
 import { heightNumber } from "../size";
-import { SystemColors } from "../system-colors";
-import { disabledCursor } from "../disabled";
 
 export const BaseButtonStyles = css`
     ${display("inline-block")} :host {
@@ -71,19 +69,6 @@ export const BaseButtonStyles = css`
     .end {
         margin-inline-start: 11px;
     }
-
-    @media (forced-colors: active) {
-        :host(.disabled),
-        :host(.disabled) .control {
-            forced-color-adjust: none;
-            background: ${SystemColors.ButtonFace};
-            border-color: ${SystemColors.GrayText};
-            color: ${SystemColors.GrayText};
-            cursor: ${disabledCursor};
-            fill: ${SystemColors.GrayText};
-            opacity: 1;
-        }
-    }
 `;
 
 export const AccentButtonStyles = css`
@@ -102,30 +87,6 @@ export const AccentButtonStyles = css`
 
     :host(.accent) .control:${focusVisible} {
         box-shadow: 0 0 0 calc(var(--focus-outline-width) * 1px) inset var(--neutral-focus-inner-accent);
-    }
-
-    @media (forced-colors: active) {
-        :host(.accent) .control {
-            forced-color-adjust: none;
-            background: ${SystemColors.Highlight};
-            color: ${SystemColors.HighlightText};
-        }
-
-        :host(.accent) .control:hover {
-            background: ${SystemColors.HighlightText};
-            border-color: ${SystemColors.Highlight};
-            color: ${SystemColors.Highlight};
-        }
-
-        :host(.accent:${focusVisible}) .control {
-            border-color: ${SystemColors.ButtonText};
-            box-shadow: 0 0 0 2px ${SystemColors.HighlightText} inset;
-        }
-
-        :host(.accent.disabled) .control {
-            background: ${SystemColors.ButtonFace};
-            color: ${SystemColors.GrayText};
-        }
     }
 `;
 
@@ -201,26 +162,6 @@ export const LightweightButtonStyles = css`
         background: var(--neutral-foreground-rest);
         height: calc(var(--focus-outline-width) * 1px);
     }
-
-    @media (forced-colors: active) {
-        :host(.lightweight) .control:hover {
-            forced-color-adjust: none;
-            color: ${SystemColors.Highlight};
-        }
-
-        :host(.lightweight) .control:hover .content::before {
-            background: ${SystemColors.Highlight};
-        }
-
-        :host(.lightweight.disabled) .control {
-            forced-color-adjust: none;
-            color: ${SystemColors.GrayText};
-        }
-    
-        :host(.lightweight.disabled) .control:hover .content::before {
-            background: none;
-        }
-    }
 `;
 
 export const OutlineButtonStyles = css`
@@ -236,12 +177,6 @@ export const OutlineButtonStyles = css`
     :host(.outline) .control:active {
         border-color: var(--neutral-outline-active);
     }
-
-    @media (forced-colors: active) {
-        :host(.outline.disabled) .control {
-            border-color: ${SystemColors.GrayText};
-        }
-    }
 `;
 
 export const StealthButtonStyles = css`
@@ -255,12 +190,5 @@ export const StealthButtonStyles = css`
 
     :host(.stealth) .control:active {
         background: var(--neutral-fill-stealth-active);
-    }
-
-    @media (forced-colors: active) {
-        :host(.stealth.disabled) .control {
-            background: none;
-            border-color: transparent;
-        }
     }
 `;
