+++ packages/web-components/fast-components/src/tabs/tab/tab.styles.ts
@@ -8,7 +8,7 @@ import {
     neutralForegroundRestBehavior,
 } from "../../styles/recipes";
 import { SystemColors } from "../../styles/system-colors";
-import { forcedColorsStylesheetBehavior } from "../styles/match-media-stylesheet-behavior";
+import { forcedColorsStylesheetBehavior } from "../../styles/match-media-stylesheet-behavior";
 
 export const TabStyles = css`
     ${display("inline-flex")} :host {
